Moved the map component to its own file.
